refactor: modified the task controller to return the tasks related to the signed in user  only.

+++ src/tasks/tasks.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, ForbiddenException } from '@nestjs/common';
 import { Task } from './task.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
+import { User } from 'src/auth/user.entity';
 
 
 @Injectable()
@@ -13,32 +14,36 @@ export class TasksService {
         @InjectRepository(TaskRepository)
         private taskRepository:TaskRepository,
     ){}
-    async getTasks(filterDto:GetTasksFilterDto):Promise<Task[]>{
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto:GetTasksFilterDto,user:User):Promise<Task[]>{
+        return this.taskRepository.getTasks(filterDto,user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(id: number,user:User): Promise<Task> {
+        const task = await this.taskRepository.findOne({where:{id,userId:user.id}});
 
         if (!task) {
             throw new NotFoundException(`Task with id:${id} not found`);
         }
+        // if(task.userId != user.id){
+        //     throw new ForbiddenException();
+        // }
+
         return task;
     }
 
-   async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return  this.taskRepository.createTask(createTaskDto);
+   async createTask(createTaskDto: CreateTaskDto,user:User): Promise<Task> {
+        return  this.taskRepository.createTask(createTaskDto,user);
 
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number,user:User): Promise<void> {
+        const result = await this.taskRepository.delete({id,userId:user.id});
         if (!result.affected) {
             throw new NotFoundException(`Task with id:${id} not found`);
         }
     }
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus,user:User): Promise<Task> {
+        const task = await this.getTaskById(id,user);
         task.status = status;
         await task.save();
         return task;
